Fix strange bug with Firefox, Vivaldi, and Skype
After running for some time, Firefox, Vivaldi and Skype seem to lose its
"content". The area that should be its window is now empty, showing what was
there before switch to the current window. When switching workspaces, it seems
that area never switches. The fix (not sure if it is a fix), is to always enable
borders.